% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{baseline_correction}
\alias{baseline_correction}
\title{Eliminate background signal and artifacts across a whole experiment}
\usage{
baseline_correction(
  raw_data,
  smooth = TRUE,
  coupled_channels,
  channels,
  pc_wells,
  thresholds
)
}
\arguments{
\item{raw_data}{The raw fluorescence data from QIAcuity}

\item{smooth}{Boolean; default:TRUE. Whether or not the function should search for break points in the data and eliminate any artifacts that skew fluorescence intensities.}

\item{coupled_channels}{Defines which channels are highly correlated.}

\item{channels}{Defines which columns contain fluorescence data.}

\item{pc_wells}{Defines which wells are the positive controls}
}
\value{
Returns data in the same format as raw_data, with background signal and artifacts removed from data.
}
\description{
This function eliminates artifacts and background signal from QIAcuity data from a whole experiment in multiple steps.
}
