% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R
\name{classify_peaks}
\alias{classify_peaks}
\title{Classify peaks previously detected by find_turnpoints()}
\usage{
classify_peaks(
  turnpoints,
  intensities,
  variable,
  channel_maxima,
  reference_peaks = NULL
)
}
\arguments{
\item{data}{Test}
}
\value{
A dataframe with turnpoints
}
\description{
This function classifies turnpoints in a distribution
}
