% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{competition_correction}
\alias{competition_correction}
\title{Perform competition correction}
\usage{
competition_correction(
  crosstalk_corrected,
  channels,
  pc_wells,
  thresholds,
  step
)
}
\arguments{
\item{crosstalk_corrected}{The crosstalk corrected fluorescence data from QIAcuity}

\item{channels}{Defines which columns in the data contain fluorescence data}

\item{pc_wells}{Wells that contain the positive controls}

\item{thresholds}{Data frame with the thresholds for all of the channels}
}
\value{
Returns a list of (1) an overview of the competition effects between the different channels and (2) the competition corrected data
}
\description{
This function takes QIAcuity fluorescence data, calculates the effects of competition between the different reactions and corrects for them
}
