% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{crosstalk_correction}
\alias{crosstalk_correction}
\title{Perform crosstalk correction}
\usage{
crosstalk_correction(
  baseline_data,
  baseline_data_dichot,
  pc_wells,
  coupled_channels,
  channels,
  thresholds
)
}
\arguments{
\item{baseline_data}{The baseline corrected fluorescence data from QIAcuity}

\item{baseline_data_dichot}{The baseline corrected fluorescence data from QIAcuity, dichotomized}

\item{pc_wells}{Wells that contain the positive controls}

\item{coupled_channels}{Defines which channels are highly correlated.}

\item{channels}{Defines which columns in the data contain fluorescence data}

\item{thresholds}{Data frame with the thresholds for all of the channels}
}
\value{
Returns a list of (1) the crosstalk corrected data and (2) an overview of the crosstalk effects between the different channels
}
\description{
This function takes QIAcuity fluorescence data, calculates the effects of crosstalk between the channels and corrects for them
}
