% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R
\name{density_threshold}
\alias{density_threshold}
\title{Define a threshold based on mixture modeling}
\usage{
density_threshold(
  input_data,
  variable,
  min_dist = 0.2,
  references = NULL,
  pc_data = NULL
)
}
\arguments{
\item{variable}{Name of the variable in the dataframe to be pulled}

\item{data}{Either an object of class dataframe (if parameter variable is defined), of a vector that can be forced to numeric}

\item{stringent}{If set to TRUE, forces script to set threshold half-way between mu of the two normal distributions. This decreases the number of false positive results.}
}
\value{
An object of class dataframe of the RFU values from the .csv-file
}
\description{
This function defines a threshold that best divides a population of data points into two separate groups based on mixture modeling.
}
