% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R
\name{find_turnpoints}
\alias{find_turnpoints}
\title{Find turnpoints in a distribution}
\usage{
find_turnpoints(data, variable = NULL)
}
\arguments{
\item{data}{Test}
}
\value{
A dataframe with turnpoints
}
\description{
This function detects turnpoints in a distribution
}
