% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R
\name{get_quantile}
\alias{get_quantile}
\title{Get quantile}
\usage{
get_quantile(input, probs = 0.25)
}
\arguments{
\item{input}{Input vector of data that can be forced to numeric.}

\item{probs}{quantile to be returned}
}
\value{
A numeric value.
}
\description{
This function gives just the numeric value of a defined quantile in a vector of data.
}
