% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R
\name{read.data.QIAcuity}
\alias{read.data.QIAcuity}
\title{Read QIAcuity data}
\usage{
read.data.QIAcuity(filename)
}
\arguments{
\item{filename}{Path to the location of the QIAcuity .csv-file to be read.}
}
\value{
An object of class dataframe of the RFU values from the .csv-file
}
\description{
This function reads in raw data .csv-files from the QIAcuity system
}
