% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{recalculate_thresholds}
\alias{recalculate_thresholds}
\title{Calculate thresholds for input data and dichotomize it}
\usage{
recalculate_thresholds(data, thresholds, step, coupled_channels)
}
\arguments{
\item{data}{The fluorescence data from QIAcuity}

\item{thresholds}{Data frame where threshold values are stored.}

\item{step}{Column of the data frame where the data should be stored.}

\item{coupled_channels}{Defines which channels are highly correlated.}
}
\value{
Returns a list of (1) the thresholds and (2) the data, transformed to 0 or 1 depending on whether the fluorescence value exceeds the threshold
}
\description{
This function takes QIAcuity fluorescence data, calculates thresholds for all channels and then classifies each partition based on these thresholds as 0 or 1.
}
