% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{saveQIAcuityResult}
\alias{saveQIAcuityResult}
\title{Save output of QIAcuity Analyisis}
\usage{
saveQIAcuityResult(
  results,
  output_path,
  scatterplots_2d = TRUE,
  scatterplots_1d = TRUE,
  save_raw_data = TRUE,
  save_baseline_corr = TRUE,
  save_crosstalk_corr = TRUE,
  save_competition_corr = TRUE,
  save_fluorescence_data = TRUE
)
}
\arguments{
\item{results}{A list of dataframes as returned by the function QIAcuityAnalysis}

\item{output_path}{path to where files and plots should be saved. Will create directory for the experiment there.}

\item{scatterplots_2d}{Set to TRUE to generate 2D scatterplots at every selected correction step. Default: TRUE}

\item{scatterplots_1d}{Set to TRUE to generate 1D scatterplots at every selected correction step. Default: TRUE}

\item{save_raw_data}{Whether plots and summaries based on raw data should be generated and saved. Default: TRUE}

\item{save_baseline_corr}{Whether plots and summaries based on baseline corrected data should be generated and saved. Default: TRUE}

\item{save_crosstalk_corr}{Whether plots and summaries based on crosstalk corrected data should be generated and saved. Default: TRUE}

\item{save_competition_corr}{Whether plots and summaries based on competition corrected data should be generated and saved. Default: TRUE}

\item{save_fluorescence_data}{Whether fluorescence data should be saved for all selected correction steps. Size: up to 70 MB per file. Default: TRUE}
}
\value{
NULL. Generates output files in the selected directory:

\strong{Result summary:}  [experiment name]\_ result\_[correction step]\_corrected.xlsx \if{html}{\out{<br />}}
Contains summary of dPCR results at this correction step. Columns: \if{html}{\out{<br />}}
\strong{Sample:} Sample ID \if{html}{\out{<br />}}
\strong{Well:} Well on plate \if{html}{\out{<br />}}
\strong{Channel:} Data from this channel is summarized in this line \if{html}{\out{<br />}}
\strong{valid:} Number of valid partitions in this well in this channel \if{html}{\out{<br />}}
\strong{negative:} Number of negative partitions in this well in this channel \if{html}{\out{<br />}}
\strong{positive:} Number of positive partitions in this well in this channel \if{html}{\out{<br />}}
\strong{CI:} 95\% confidence interval of the quantification \if{html}{\out{<br />}}
\strong{poisson_corrected_targets:} Total number of copies of the DNA target in this channel in this well \if{html}{\out{<br />}}
\strong{mean_z_score:} Based on a positive sample. Z-score quantifies, by how many standard deviations the fluorescence intensity of a partition deviates from the mean fluorescence intensity of positive partitions in a positive control. Values lower than -2 usually indicate that a partition is background noise, while values greater than 2 can indicate imaging artifacts. \if{html}{\out{<br />}}
\strong{Warning:} Warnings are printed if there is any indication of artifacts or uncertain quantification \if{html}{\out{<br />}}

\strong{2d scatterplots:} [experiment name]\_ [correction step]\_[channel x axis][channel y axis][channel color].png

\strong{1d scatterplots:} [experiment name]\_ [correction step]\_[channel].png

\strong{thresholds:} [experiment name]\_ thresholds.xlsx

\strong{crosstalk analysis:} [experiment name]\_ crosstalk\_analysis.xlsx

\strong{competition analysis:} [experiment name]\_ competition\_analysis.xlsx

\strong{Signal-to-noise ratio analysis:} [experiment name]\\emph{snr}[correction step].xlsx

\strong{Fluorescence data:} [experiment name]\_[correction step].csv \if{html}{\out{<br />}}
Contains the fluorescence data. Each partition has a defined position on the plate \(overall partition ID, Well and ID of the partition within the well\), sample name, and fluorescence intensities in all the channels.
}
\description{
This function summarizes and visualizes the results of one or multiple QIAcuity experiments.
}
