% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R, R/core_functions.R
\name{setup}
\alias{setup}
\title{Load necessary packages}
\usage{
setup()

setup()
}
\description{
This function loads all necessary packages for the script to run.

This function loads all necessary packages for the script to run.
}
