% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{smooth_data}
\alias{smooth_data}
\title{Eliminate background signal and artifacts}
\usage{
smooth_data(data, current_channel, upper_lim, lower_lim)
}
\arguments{
\item{data}{The fluorescence data from QIAcuity}

\item{current_channel}{Which column the artifact correction should be performed on.}

\item{upper_lim}{Upper RFU limit of which partitions should be considered as part of the negative group}

\item{lower_lim}{Lower RFU limit of which partitions should be considered as part of the negative group}
}
\value{
Returns data in the same format, with only the data in column current_channel altered to remove artifacts
}
\description{
This function eliminates artifacts by searching for break points where sudden changes in fluorescence intensity on the plate occur, and fits a model to each region identified to estimate artifact effects and subtract them from the data.
}
