% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R
\name{classify_peaks}
\alias{classify_peaks}
\title{Classify peaks previously detected by find_turnpoints()}
\usage{
classify_peaks(
  turnpoints,
  intensities,
  variable,
  channel_maxima,
  reference_peaks = NULL
)
}
\arguments{
\item{turnpoints}{Output of find_turnpoints}

\item{intensities}{Fluorescence intensity values in a data frame}

\item{variable}{Which column in intensities is the channel to be examined}

\item{channel_maxima}{A named list giving the maximum values expected for each of the channels found in intensities}

\item{reference_peaks}{A dataframe of previously classified peaks to guide this classification; optional.}
}
\value{
A dataframe with turnpoints
}
\description{
This function classifies turnpoints in a distribution
}
