% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R
\name{density_threshold}
\alias{density_threshold}
\title{Define a threshold based on mixture modeling}
\usage{
density_threshold(
  input_data,
  variable,
  min_dist = 0.2,
  references = NULL,
  pc_data = NULL
)
}
\arguments{
\item{input_data}{An object of class dataframe containing fluorescence intensities fom ddPCR partitions.}

\item{variable}{Name of the variable in the dataframe to be pulled}

\item{min_dist}{Determines the minimum distance between threshold and negative partitions, always as a fraction of the distance between true positive and true negative populations.}

\item{references}{Reference peaks for guiding classification}

\item{pc_data}{Positive control data as a reference to calculate maximum expected values for each channel}
}
\value{
A numeric value giving the optimal threshold for distinguishing the true positive and true negative populations in the data.
}
\description{
This function defines a threshold that best divides a population of data points into two separate groups based on mixture modeling.
}
