% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R, R/peak_functions.R
\name{find_turnpoints}
\alias{find_turnpoints}
\title{Find turnpoints in a distribution}
\usage{
find_turnpoints(data, variable, n_points = 30)

find_turnpoints(data, variable, n_points = 30)
}
\arguments{
\item{data}{A dataframe or matrix only with numeric RFU values. Columns: Channels, Rows: partitions. Important:}

\item{variable}{A character specifying the column of data to be examined}
}
\value{
A dataframe with turnpoints

A dataframe with turnpoints
}
\description{
This function detects turnpoints in a distribution

This function detects turnpoints in a distribution
}
