% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIAcuity_package_V3.R
\name{QIAcuityAnalysis}
\alias{QIAcuityAnalysis}
\title{Analyze output of a QIAcuity experiment}
\usage{
QIAcuityAnalysis(
  input_path,
  output_path,
  coupled_channels = data.frame(ch1 = character(length = 0), ch2 = character(length =
    0))
)
}
\arguments{
\item{input_path}{path to where the .zip or .csv files exported from the QIAcuity software are saved. No files are written here. The function can handle multiple experiments saved in the same directory. It will always analyze all of them and create separate results for them.}

\item{output_path}{path to where any results and plots should be saved. Directory for each experiment will be created there.}

\item{coupled_channels}{Data frame containing any channels the targets of which are on the same chromosome or otherwise to be expecte in the same partitions}
}
\value{
A list containing a list of dataframes for each experiment.
}
\description{
This function takes the files exported from the QIAcuity software for one or more experiments and analyzes them in an automated way.
}
\details{
This function returns a list with an element for each experiment that was in the directory set by path. In each element there are the following data frames: The data for all channels in raw data, after baseline correction, cross talk correction and competition correction; cross talk and competition analyses that give an estimate of the magnitude of the effects the different reactions have on each other; thresholds gives the thresholds for each channel at each step; Volumes contains information about the total volume contained in the partitions of each well.
}
